Add validators to CRUD operations and test them successfully

+++ example-app/app/Http/Controllers/UserController.php
@@ -5,10 +5,27 @@ namespace App\Http\Controllers;
 use Illuminate\Http\Request;
 use App\Models\User;
 use Illuminate\Support\Facades\Hash;
+use Illuminate\Support\Facades\Validator;
+use Illuminate\Validation\Rule;
 
 class UserController extends Controller
 {
     public function addUser(Request $request) {
+
+        $rules = [
+            'name' => 'required|string|max:255',
+            'email' => 'required|email|unique:users,email|max:255',
+            'password' => 'required|string|min:8',
+        ];
+
+        $validator = Validator::make($request->all(), $rules);
+
+        if ($validator->fails()) {
+            return response()->json([
+                'error' => $validator->errors(),
+            ], 400);
+        }
+
         $user = new User;
         $name = $request->input('name');
         $email = $request->input('email');
@@ -25,7 +42,22 @@ class UserController extends Controller
     }
 
     public function getUser(Request $request, $id) {
+
+        $validator = Validator::make(['id' => $id], ['id' => 'required|numeric',]);
+
+        if ($validator->fails()) {
+            return response()->json([
+                'error' => 'Invalid user ID format',
+            ], 400);
+        }
+
         $user = User::find($id);
+
+        if (!$user) {
+            return response()->json([
+                'error' => 'User not found!',
+            ], 404);
+        }
 
         return response()->json([
             'message' => 'Retrieved user successfully!',
@@ -34,7 +66,37 @@ class UserController extends Controller
     }
 
     public function editUser(Request $request, $id) {
+
+        $validator = Validator::make(['id' => $id], ['id' => 'required|numeric',]);
+
+        if ($validator->fails()) {
+            return response()->json([
+                'error' => 'Invalid user ID format',
+            ], 400);
+        }
+
         $user = User::find($id);
+
+        if (!$user) {
+            return response()->json([
+                'error' => 'User not found!',
+            ], 404);
+        }
+
+        $rules = [
+            'name' => 'required|string|max:255',
+            'email' => 'required|email|unique:users,email|max:255',
+            'password' => 'required|string|min:8',
+        ];
+
+        $validator = Validator::make($request->all(), $rules);
+
+        if ($validator->fails()) {
+            return response()->json([
+                'error' => $validator->errors(),
+            ], 400);
+        }
+
         $inputs = $request->except('_method');
         $user->update($inputs);
 
@@ -45,7 +107,23 @@ class UserController extends Controller
     }
 
     public function deleteUser(Request $request, $id) {
+
+        $validator = Validator::make(['id' => $id], ['id' => 'required|numeric',]);
+
+        if ($validator->fails()) {
+            return response()->json([
+                'error' => 'Invalid user ID format',
+            ], 400);
+        }
+
         $user = User::find($id);
+
+        if (!$user) {
+            return response()->json([
+                'error' => 'User not found!',
+            ], 404);
+        }
+
         $user->delete();
 
         return response()->json([
